// frontend/src/app/api/models/pelada-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ParticipanteResponse } from '../models/participante-response';
import { UsuarioResumoResponse } from '../models/usuario-resumo-response';
export interface PeladaResponse {
  cidade: string;
  data: string;
  descricao?: string;
  endereco: string;
  estado: string;
  horaFim: string;
  horaInicio: string;
  id: number;
  localNome: string;
  maxParticipantes: number;
  nome: string;
  organizador: UsuarioResumoResponse;
  participantes: Array<ParticipanteResponse>;
  status: 'AGENDADA' | 'CONFIRMADA' | 'EM_ANDAMENTO' | 'FINALIZADA' | 'CANCELADA';
  statusDescricao: string;
  tipoCampo: 'FUTSAL' | 'SOCIETY';
  tipoCampoDescricao: string;
  totalConfirmados: number;
  vagasRestantes: number;
  valorPorJogador?: number;
}